Design Files and Pictures
The main design file is in sketchup 2016 format (.skp), which is the shipping container greenhouse system.  It's missing some key parts such as the LED lighting system, plumbing, working area to process trays, storage, etc.  It's a rough sketch by engineering standards, but it's a good start.  I have a lot of these things figured out mentally, which is where I live most of the time.